int result changed to double hence all the program is rewritten

+++ src/main/java/core/basesyntax/Calculable.java
@@ -1,5 +1,5 @@
 package core.basesyntax;
 
 public interface Calculable {
-    int calculate(int a, int b, char operand);
+    double calculate(int a, int b, char operand);
 }
